Image processing and custom model use
This update automatically scans image directory for all jpg images which are then used for pix-plot analysis. I also added a model_dir gooey option that allows you to specify the location of a directory where your custom model is held. Another points the program to the specific model within that directory.